requirements.txt
# ==========================================
# AI 时间追踪系统 - 依赖清单 v2.0
# ==========================================

# --- Web 界面 ---
streamlit>=1.28.0       # WebUI 框架
plotly>=5.18.0          # 交互式图表
pandas>=2.0.0           # 数据处理

# --- 系统监控 (Windows) ---
psutil>=5.9.0           # 进程信息获取
pywin32>=306            # Windows API (win32gui, win32process)
uiautomation>=2.0.18    # UI自动化 (获取浏览器URL)
pynput>=1.7.6           # 全局键鼠监听

# --- 系统托盘 ---
pystray>=0.19.5         # 系统托盘图标
Pillow>=10.0.0          # 图像处理 (托盘图标绘制)

# --- AI API ---
openai>=1.3.0           # OpenAI 兼容 API 客户端
httpx>=0.25.0           # HTTP 客户端 (openai 依赖)

# --- 可选依赖 ---
# watchdog>=3.0.0       # 文件监控 (配置热重载)
# schedule>=1.2.0       # 定时任务 (定期报告)
